requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg2-binary
python-jose[cryptography]
